+++ src/web/client/App.tsx
@@ -1382,15 +1382,15 @@ export function App({ initialTourId, replyAgent }: AppProps): React.JSX.Element 
           </span>
         </div>
         <div className="tour-header-right">
+          <TourStatsIndicator
+            additions={tourStats.additions}
+            deletions={tourStats.deletions}
+          />
           <SequencePill
             idx={pillIdx}
             total={view.nav.navTotal}
             onPrev={() => navigateBy(-1)}
             onNext={() => navigateBy(1)}
-          />
-          <TourStatsIndicator
-            additions={tourStats.additions}
-            deletions={tourStats.deletions}
           />
           <LayoutToggle layout={layout} onChange={setLayoutChoice} />
         </div>